Optimized algorithm, refactored tests and added view and encoding scheme (#2)
